desktop-agent/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1

# --- GUI 依赖 ---
pystray>=0.19.0
Pillow>=10.0.0
# ttkbootstrap 可选，提供现代化 tkinter 界面
ttkbootstrap>=1.10.0
